Convert IDs to Snowflake-like IDs
Closes #2

+++ migrations/10.sql
@@ -1,6 +1,6 @@
 CREATE TABLE tournament_staff(
 	id INT NOT NULL,
-	tournament INT NOT NULL,
+	tournament BIGINT NOT NULL,
 	privileges INT NOT NULL,
 	PRIMARY KEY(id, tournament)
 );
