+++ migrations/2.sql
@@ -1,5 +1,5 @@
 CREATE TABLE tournament_rules(
-	id INT NOT NULL,
+	id BIGINT NOT NULL,
 	rules TEXT NOT NULL,
 	FOREIGN KEY (id) REFERENCES tournaments (id)
 		ON DELETE CASCADE
